+++ RSPiX/Src/BLUE/unix/Bsound.cpp
@@ -51,15 +51,15 @@ static void sdl_audio_callback(void *userdata, Uint8 *stream, int len)
 
 static bool audio_opened = false;
 static SDL_AudioSpec desired;
-static long cur_buf_time = 0;
-static long max_buf_time = 0;
+static int32_t cur_buf_time = 0;
+static int32_t max_buf_time = 0;
 
 extern int16_t rspSetSoundOutMode(				// Returns 0 if successfull, non-zero otherwise
-	long lSampleRate,								// In:  Sample rate
-	long lBitsPerSample,							// In:  Bits per sample
-	long lChannels,								// In:  Channels (mono = 1, stereo = 2)
-	long lCurBufferTime,							// In:  Current buffer time (in ms.)
-	long lMaxBufferTime,							// In:  Maximum buffer time (in ms.)
+	int32_t lSampleRate,								// In:  Sample rate
+	int32_t lBitsPerSample,							// In:  Bits per sample
+	int32_t lChannels,								// In:  Channels (mono = 1, stereo = 2)
+	int32_t lCurBufferTime,							// In:  Current buffer time (in ms.)
+	int32_t lMaxBufferTime,							// In:  Maximum buffer time (in ms.)
 	RSP_SND_CALLBACK callback,					// In:  Callback function
 	ULONG ulUser)									// In:  User-defined value to pass to callback
 {
@@ -131,11 +131,11 @@ extern int16_t rspSetSoundOutMode(				// Returns 0 if successfull, non-zero othe
 }
 
 extern int16_t rspGetSoundOutMode(				// Returns 0 if successfull, non-zero otherwise
-	long* plSampleRate,							// Out: Sample rate or -1 (unless NULL)
-	long* plBitsPerSample,				// Out: Bits per sample or -1 (unless NULL)
-	long* plChannels,					// Out: Channels (mono=1, stereo=2) or -1 (unless NULL)
-	long* plCurBufferTime,				// Out: Current buffer time or -1 (unless NULL)
-	long* plMaxBufferTime)			// Out: Maximum buffer time or -1 (unless NULL)
+	int32_t* plSampleRate,							// Out: Sample rate or -1 (unless NULL)
+	int32_t* plBitsPerSample,				// Out: Bits per sample or -1 (unless NULL)
+	int32_t* plChannels,					// Out: Channels (mono=1, stereo=2) or -1 (unless NULL)
+	int32_t* plCurBufferTime,				// Out: Current buffer time or -1 (unless NULL)
+	int32_t* plMaxBufferTime)			// Out: Maximum buffer time or -1 (unless NULL)
 {
     SET(plSampleRate, desired.freq);
     SET(plBitsPerSample, desired.format & 0x00FF);
@@ -146,7 +146,7 @@ extern int16_t rspGetSoundOutMode(				// Returns 0 if successfull, non-zero othe
 }
 
 extern void rspSetSoundOutBufferTime(
-	long lCurBufferTime)						// In:  New buffer time
+	int32_t lCurBufferTime)						// In:  New buffer time
 {
     cur_buf_time = lCurBufferTime;
 }
@@ -193,17 +193,17 @@ extern int16_t rspIsSoundOutPaused(void)	// Returns TRUE if paused, FALSE otherw
     return((SDL_GetAudioStatus() == SDL_AUDIO_PAUSED) ? TRUE : FALSE);
 }
 
-extern long rspGetSoundOutPos(void)		// Returns sound output position in bytes
+extern int32_t rspGetSoundOutPos(void)		// Returns sound output position in bytes
 {
     return 0;
 }
 
-extern long rspGetSoundOutTime(void)		// Returns sound output position in time
+extern int32_t rspGetSoundOutTime(void)		// Returns sound output position in time
 {
     return 0;
 }
 
-extern long rspDoSound(void)
+extern int32_t rspDoSound(void)
 {
     // no-op; in Soviet Russia, audio callback pumps YOU.
     //  seriously, the SDL audio callback runs in a seperate thread.
